\name{cdprompt}
\alias{cdprompt}
\title{Support routine for cd-organized workspace hierarchy.}
\description{Sets the command-line prompt to the correct value (see \code{\link{cd}}, and the notes on the option \code{abbreviate.cdprompt}); useful if the prompt somehow becomes corrupted. \code{cdprompt} never seems necessary in R{} but has been useful in the S+ manifestations of \code{mvbutils}, where system bugs are commoner.
}
\usage{
cdprompt()
}
\seealso{\code{\link{cd}}
}
\examples{
cdprompt()
}
\author{Mark Bravington}
\keyword{utilities
}
