\name{unpackage}
\alias{unpackage}
\title{Convert existing source package into task package}
\description{
Converts an existing source package into a task package. A subdirectory with the package name will be created under the current working directory, and will be populated with a ".RData" file and various other files/directories from the source package. All Rd files will be turned into flat-format help in the ".RData", either attached to functions or as stand-alone "*.doc" text objects, as per \code{\link{help2flatdoc}}. The subdirectory will also be made into a \emph{task}, i.e. it will be added to the "tasks" vector in the current workspace that \code{\link{cd}} uses to keep track of the task hierarchy.

Roxygen pre-function comments are by default removed; see \code{keep_Roxygen} argument.
}
\usage{
unpackage(spath, force = FALSE, alias=x,
    drop_Conly=TRUE, trust_importFrom=FALSE,
    zap_Cloaders= TRUE, keep_Roxygen= FALSE)
}
\arguments{
\item{ spath}{where to find the source package}
\item{ force}{if TRUE, overwrite any previous contents of task package without prompting.}
\item{ alias}{if you \emph{don't} want the package to go into a folder (and task) with its obvious name, then set eg \code{alias="temp_mypack"}; useful for comparing versions.}
\item{ drop_Conly}{?exclude pure R{} wrappers of dot-Call/dot-External? Normally (I presume) those are auto-generated, and will be regenerated by \code{pre.install} in the "Cloaders" business, so this \emph{should} be OK.}
\item{ trust_importFrom}{Whether to force-copy all importFrom() NAMESPACE directives into the maintained package. Used to be TRUE by default, but I think FALSE is better...}
\item{ zap_Cloaders}{whether to remove \code{run_Cloaders_<mypack>}, if it's present; though it shouldn't really be, I think, cos it seems to prevent installation. (It's regenerated by \code{pre.install} automatically.) IOW, \code{run_Cloaders_<mypack>} should be removed before a git push unless you are using \code{mvbutils} package building system.}
\item{ keep_Roxygen}{I thoroughly dislike Roxygen, and my pref is to generate the help for each function--- stored in its \code{docattr} attribute--- directly from the dot-Rd file. If anyone ever really wants Roxygen, they can try \code{Rd2Roxygen} (beware the bugs...). So my default is FALSE, but if you set TRUE then \code{unpackage} will make a big effort to store the original Roxybollocks.}
}
\details{
The NAMESPACE file won't be copied; instead, it will be auto-generated by \code{\link{pre.install}}. Therefore, some features of the original NAMESPACE may be lost. You can either copy the NAMESPACE manually (in which case, you'll need to maintain it by hand), or write a "pre.install.hook.MYPACK" function.

The DESCRIPTION file becomes a character vector called e.g. \code{<mypack>.DESCRIPTION}, of class "cat" (see \code{\link{as.cat}}).

Functions in the ".RData" may be saved with extra attributes, in particular "doc" (deduced from a dot-Rd file) but perhaps other things too that they acquire after the package code is \code{source}d. Attributes that are character vector will acquire class "docattr", so that they won't be fully displayed during default printing of the function; to see them, use e.g. \code{as.cat( attr( myfun, "myatt"))} or \code{unclass( attr( myfun, "myatt"))} or, if you are using the \pkg{atease} package, just \code{as.cat( myfun@myatt)}. Editing the function with \code{\link{fixr}} will display the character attributes in full.

Any environment objects found in the package's environment (its namespace environment) will be dropped from the ".RData" file, with a warning; this is to avoid dramas on reloading.
}
\seealso{\code{\link{pre.install}}, \code{\link{mvbutils.packaging.tools}}
}
\keyword{misc}
