\name{set.presave.hook.mvb}
\alias{set.presave.hook.mvb}
\title{Do something-or-other}
\description{I have forgotten what this function is for, but probably the only reason it's documented is to make sure it's exported... My advice: don't use it!
}
\usage{
set.presave.hook.mvb( hook, set=TRUE)
}
\arguments{
\item{ hook}{can't remember}
\item{ set}{boolean...}
}
\keyword{internal
}
